requirements.txt
Pyserial
keyboard